# MM-64209: Optimize completePopulatingCategoryChannelsT for MySQL