@agnivade 2025-05-09
/cherry-pick release-10.5